pyproject.toml
[project]
name = "course-assistant-agent"
version = "0.1.0"
description = "培养方案缺口分析智能体"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.0.26",
    "langchain>=0.1.0",
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "streamlit>=1.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
